requirements.txt
streamlit==1.52.2
pandas ==2.3.3
pytest==9.0.2
pydentic=2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
requests==2.32.5
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
mysql-connector-python==9.5.0
uvicorn==0.40.0
